+++ Server/main.js
@@ -3,13 +3,12 @@ const cors = require('cors');
 const app = express();
 require('dotenv').config();
 require('./config/database');
-const departmentRoute = require("./routers/departmentRouter.js")
+const storeRoute= require("./routers/storeRouter.js")
 
 app.use(cors());
 app.use(express.json({ limit: '30mb', extended: true }));
 app.use(express.urlencoded({ limit: '30mb', extended: true }));
-
-app.use('/department',departmentRoute)
+app.use('/store',storeRoute)
 
 let port = 8000;
 app.listen(port, () => {
