+++ client/src/compontes/stores/stores.jsx
@@ -1,12 +1,19 @@
 import React from 'react';
+import DesktopNav from "../navbar/DesktopNav"
+import MobileNav from "../navbar/MobileNav"
+import Store from './store';
 import './style.css';
 import vegetablesImage from '../../assets/images/Screenshot 2022-12-14 232104.png';
 import smallHome from '../../assets/images/smallhome.png';
 import businessLiaisonLogo from '../../assets/images/businessLiaisonLogo.png';
 import onion from '../../assets/images/products_images/onions.jpg';
-function StoreList() {
+
+
+export default function stores() {
   return (
     <>
+    <DesktopNav />
+    <MobileNav />
       <div className="background-image w-full">
         <img src={vegetablesImage} alt="" width="100%" className="shadow-xl" />
       </div>
@@ -66,86 +73,10 @@ function StoreList() {
             </p>
           </div>
           <div className="flex flex-col gap-1">
-            <div className="flex flex-col sm:flex-row items-center sm:items-end border w-full gap-4 p-2">
-              <img src={smallHome} alt="" />
-              <div>
-                <h4>חוות הבית</h4>
-                <p></p>
-                <p>חנות ירקות חוות הבית במודעין-מכבים-רעות מגישה</p>
-                <div className="flex">
-                  <img
-                    className="business-liaison-logo"
-                    src={businessLiaisonLogo}
-                    alt=""
-                  />
-                  <div>
-                    <p>עמק זבולון 3,</p>
-                    <p>מודעין, אזור המרכז</p>
-                  </div>
-                </div>
-              </div>
-              <button className="store-btn">קנה בחנות זו</button>
-            </div>
-            <div className="flex flex-col sm:flex-row items-center sm:items-end border w-full gap-4 p-2">
-              <img src={smallHome} alt="" />
-              <div>
-                <h4>חוות הבית</h4>
-                <p></p>
-                <p>חנות ירקות חוות הבית במודעין-מכבים-רעות מגישה</p>
-                <div className="flex">
-                  <img
-                    className="business-liaison-logo"
-                    src={businessLiaisonLogo}
-                    alt=""
-                  />
-                  <div>
-                    <p>עמק זבולון 3,</p>
-                    <p>מודעין, אזור המרכז</p>
-                  </div>
-                </div>
-              </div>
-              <button className="store-btn">קנה בחנות זו</button>
-            </div>
-            <div className="flex flex-col sm:flex-row items-center sm:items-end border w-full gap-4 p-2">
-              <img src={smallHome} alt="" />
-              <div>
-                <h4>חוות הבית</h4>
-                <p></p>
-                <p>חנות ירקות חוות הבית במודעין-מכבים-רעות מגישה</p>
-                <div className="flex">
-                  <img
-                    className="business-liaison-logo"
-                    src={businessLiaisonLogo}
-                    alt=""
-                  />
-                  <div>
-                    <p>עמק זבולון 3,</p>
-                    <p>מודעין, אזור המרכז</p>
-                  </div>
-                </div>
-              </div>
-              <button className="store-btn">קנה בחנות זו</button>
-            </div>
-            <div className="flex flex-col sm:flex-row items-center sm:items-end border w-full gap-4 p-2">
-              <img src={smallHome} alt="" />
-              <div>
-                <h4>חוות הבית</h4>
-                <p></p>
-                <p>חנות ירקות חוות הבית במודעין-מכבים-רעות מגישה</p>
-                <div className="flex">
-                  <img
-                    className="business-liaison-logo"
-                    src={businessLiaisonLogo}
-                    alt=""
-                  />
-                  <div>
-                    <p>עמק זבולון 3,</p>
-                    <p>מודעין, אזור המרכז</p>
-                  </div>
-                </div>
-              </div>
-              <button className="store-btn">קנה בחנות זו</button>
-            </div>
+            <Store smallHome={smallHome} business={businessLiaisonLogo} />
+            <Store smallHome={smallHome} business={businessLiaisonLogo} />
+            <Store smallHome={smallHome} business={businessLiaisonLogo} />
+            <Store smallHome={smallHome} business={businessLiaisonLogo} />
             <div>
               <button className="pul-btn text-gray-800">+</button>
             </div>
@@ -183,7 +114,6 @@ function StoreList() {
         </div>
       </div>
     </>
-  );
+  )
 }
 
-export default StoreList;
