--- client/src/pages/Store.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-function Store() {
-    return (
-        <div>Store</div>
-    )
-}
-
-export default Store
