+++ client/src/pages/StorePage.jsx
@@ -1,0 +1,10 @@
+import React from 'react'
+import Stores from '../compontes/stores/stores'
+
+export default function StorePage() {
+  return (
+    <div>
+        <Stores />
+    </div>
+  )
+}
